+++ backend/src/main/java/com/ssafy/home/global/config/security/WebSecurityConfiguration.java
@@ -1,7 +1,7 @@
 package com.ssafy.home.global.config.security;
 
 import com.ssafy.home.domain.user.repository.MemberRepository;
-import com.ssafy.home.global.auth.JwtTokenProvider;
+import com.ssafy.home.global.auth.jwt.JwtTokenProvider;
 import com.ssafy.home.global.auth.filter.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
